Move zabbix state tests to pytest
